updated persistence annotations

+++ src/main/java/com/abbyhowe/LearnFolio/repository/AssignmentData.java
@@ -9,26 +9,26 @@ import java.util.Map;
 public class AssignmentData {
 
     //holds static methods.
-    //this will be the place to hold students.
-    private static final Map<Integer, Assignment> assignments = new HashMap<>();
+    //this will be the place to hold assignments.
+    private static final Map<Long, Assignment> assignments = new HashMap<>();
 
-    //gets all students.
-    public static Collection<Assignment> getAll(){
+    //gets all assignments.
+    public static Collection<Assignment> getAll() {
         return assignments.values();
     }
 
-    //get a single student.
-    public static Assignment getById(int id){
+    //get a single assignment.
+    public static Assignment getById(Long id) {
         return (Assignment) assignments.values();
     }
 
-    //add student
-    public static void add(Assignment assignment){
+    //add assignment
+    public static void add(Assignment assignment) {
         assignments.put(assignment.getId(), assignment);
     }
 
-    //deletes an event
-    public static void remove(int id){
+    //deletes an assignment
+    public static void remove(Long id) {
         assignments.remove(id);
     }
 }
